Fix generation of configured-clang-sysroot-arguments for system llvm
This file is needed for CHPL_LLVM=system since the -resource-dir
might need to be set. However it was only being computed with
CHPL_LLVM=llvm with host compiler=clang. Additionally, it does not
make sense to store this file in the same place for CHPL_LLVM=system
and for CHPL_LLVM=llvm, so now the system version is installed to
 third-party/llvm/install/system/configured-clang-sysroot-arguments
(the argument being that the platform/compiler are no longer necessary
 since it refers to a clang that is installed system-wide).