Add copy-no-alias flag for now
In the first example var x = new R(0); var y = x;, the compiler is
currently considering x to be dead at end-of-block because of the
copy-initialization, for a record containing shared. For now, adjust
the analysis to allow a flag to indicate that init= does not result
in an aliasing thing.