Resolve problem with parallel/taskPar/sungeun/private
Because of the different way in which the initCopy for an iterator
was being resolved, certain functions were not visible. Resolved
that by adjusting fixInstantiationPointAndTryResolveBody to
consider the case of a type that is not instantiated.

Additionally, avoided recursion in resolving initCopy for an iterator
by adding two `pragma "no copy"`s.