Use the POI scope to start isMoreVisible
For functions/ferguson/hijacking/issue-18119-rev.chpl.  The idea is that
something can be more visible at the POI.  Before PR #8079, we had
isMoreVisible consult both regular scopes and POIs. However, after PR
there is a candidate available without consulting POI. As a result, the
candidates to be disambiguated must always come from a single POI. This
commit adjusts the compiler to track the scope of that single POI and use
it in isMoreVisible.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>