Resolves #20441
When given a regexp that matches an empty string, the iterators
`regex.matches()` supplied with an empty string and
`channel.matches()` supplied with an empty channel
now yield exactly one match.

When given such a regexp and a non-empty channel, `channel.matches()`
now mimicks the behavior of `regex.matches()` on a non-empty string.

IMPLEMENTATION NOTES

The key changes are:

* record a valid match even when `spPtr[i].data() == NULL`
  in qio_regex_match()

* restore `oldPosition` in proc _channel._extractMatch()

* execute `qio_channel_advance(1 byte)` when the matched string is empty
  in iter _channel.matches()

* minor cleanups "while there"

For the last point, we decided to insert this additional advance
into the iterator rather than into qio_regex_match() -- because
this advancement is needed when iterating. Cf. if we are just getting
an empty-string match on a channel, it should not get advanced.

Another place where this latter point comes to play is when calling
`readf` whose format string matches an empty string. In this case
the channel should not get advanced and subsequent such calls
will match at the same offset in the channel.
I added a test illustrating this.

NEXT STEPS: see #20864

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>