Add PRIM_CLEANUP_LOCAL_VARIABLE, wire to llvm.invariant.end
PRIM_CLEANUP_LOCAL_VARIABLE is added during callDestructors.  I am
concered that the code generator won't recognize the deinitialization
for a variable (in the form of either call to chpl__autoDestroy or a
deinit method) because one or both of those calls might be inlined.

Because PRIM_CLEANUP_LOCAL_VARIABLE is about a variable, I adjusted
copy propagation to avoid copy propagating with it.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>