Adjust disambiguation to prefer methods
In the language specification as well as in the production resolver.

Note that this change is unlikely to impact the production resolver,
because earlier passes are determining what is a method call and what is
a function call. However, in the future, we would like the resolver to
be responsible for this choice.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>